third_party/blink/web_tests/platform/generic/http/tests/security/preload-script-crossorigin-fails-cross-origin-expected.txt

CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-script.php?delay=100&cors=false&value=FAIL' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Preload of script with CORS failure only shows one error

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Did preload.
PASS Script error.
PASS successfullyParsed is true

TEST COMPLETE
